backend/requirements.txt
elevenlabs
python-dotenv
fastapi
uvicorn[standard]
google-auth-oauthlib
google-api-python-client
httpx
supabase
python-jose[cryptography]
python-multipart
